0425-1 1、修复分类修改bug 2、增加网关模块 3、增加用户认证模块

+++ UserService/src/main/java/com/q7w/Entity/RoleResource.java
@@ -19,10 +19,10 @@ public class RoleResource extends BaseEntity implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
 
-    private Integer rid;
+    private Long rid;
 
-    private Integer pid;
+    private Long pid;
 }
